A project developed with Python
This project was developed by Neel Bhatt for the Computer Engineering
Class TEJ4M0 assignment called cube world. Please check the repository
for the problem posed in the assignment and the algorithm for it.

Feel free to contact me regarding questions and scopes for improvement
at neel1302@gmail.com

Happy Coding!